pyproject.toml
[project]
name = "cinepyle"
version = "0.1.0"
description = "Korean cinema Telegram bot - box office rankings, new movie alerts, IMAX notifications, and nearby theater finder"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "lxml>=5.3.0",
    "playwright>=1.49.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
cinepyle = "cinepyle.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cinepyle"]
